% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{as.matrix.gpuMatrix}
\alias{as.matrix.gpuMatrix}
\title{Convert the gpuMatrix object into a matrix}
\usage{
\method{as.matrix}{gpuMatrix}(x)
}
\arguments{
\item{x}{an gpuMatrix object}
}
\description{
The function will convert the gpuMatrix object into a matrix,
if you have run any GPU functions on the gpuMatrix object,
please call \code{download(x)} to synchronize the data before calling this function.
}
