% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{as.vector.gpuMatrix}
\alias{as.vector.gpuMatrix}
\title{Convert the gpuMatrix object into a vector}
\usage{
\method{as.vector}{gpuMatrix}(x, mode = NULL)
}
\arguments{
\item{x}{an gpuMatrix object}

\item{mode}{This argument is only for compatibility. It does not take any effect.}
}
\description{
The function will convert the gpuMatrix object into a vector,
if you have run any GPU functions on the gpuMatrix object,
please call \code{download(x)} to synchronize the data before calling this function.
}
