% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuApply.R
\name{compileGPUCode}
\alias{compileGPUCode}
\title{Compile the R function without excute it in the device.}
\usage{
compileGPUCode(X, FUN, ..., .macroParms = NULL,
  .options = gpuSapply.getOption())
}
\arguments{
\item{X}{a vector that \code{FUN} will loop over.}

\item{FUN}{The function to be applied to each elemtn of \code{X}}

\item{...}{optional arguments to \code{FUN}}

\item{.macroParms}{The function argument that will be treated as macro in the code.
If an argument is treated as macro, its value cannot be changed by the code}

\item{.options}{The package and openCL compilation options, please call \code{gpuSapply.getOption()} to get all the available options}
}
\description{
Compile the R function without excute it in the device.
}
