% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\docType{methods}
\name{dim}
\alias{dim}
\title{Dimensions of an gpuMatrix object}
\usage{
\S4method{dim}{gpuMatrix}(x)
}
\arguments{
\item{x}{an gpuMatrix object}
}
\description{
Retrieve the dimension of an gpuMatrix object
}
