% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{download}
\alias{download}
\title{Get the data from the device}
\usage{
download(obj)
}
\arguments{
\item{obj}{an gpuMatrix object}
}
\description{
Get the data from the device. You should explicitly call it when you want to collect the data from the device.
}
