% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\docType{methods}
\name{[,gpuMatrix,ANY,ANY,missing-method}
\alias{[,gpuMatrix,ANY,ANY,missing-method}
\alias{[,gpuMatrix-method}
\alias{[}
\alias{[<-,gpuMatrix,ANY,ANY,numeric-method}
\title{extract/set parts of the data in gpuMatrix object}
\usage{
\S4method{[}{gpuMatrix,ANY,ANY,missing}(x, i = NA, j = NA, ...,
  drop = TRUE)

\S4method{[}{gpuMatrix,ANY,ANY,numeric}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an gpuMatrix object}

\item{i, j}{indices specifying elements to extract or replace. The index j can be missing or empty.}

\item{...}{This argument is only for compatibility. It does not take any effect.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension.}

\item{value}{The value you want to set}
}
\value{
A matrix subset

no return value
}
\description{
extract/set parts of the data in gpuMatrix object
}
\concept{Extract}
