% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuResourcesManager.R
\name{getDeviceDetail}
\alias{getDeviceDetail}
\title{Get the device detailed information, call 'getDeviceList()' first to figure out the index before use this function}
\usage{
getDeviceDetail(i)
}
\arguments{
\item{i}{The device index}
}
\description{
Get the device detailed information, call 'getDeviceList()' first to figure out the index before use this function
}
