% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{getDeviceInfo}
\alias{getDeviceInfo}
\title{Get the ith device information, call 'getDeviceList()' first to figure out the index before use this function}
\usage{
getDeviceInfo(i)
}
\arguments{
\item{i}{numeric The device index}
}
\description{
Get the ith device information, call 'getDeviceList()' first to figure out the index before use this function
}
