% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{getDeviceList}
\alias{getDeviceList}
\title{The function is used to obtain all the opencl-enable devices}
\usage{
getDeviceList()
}
\description{
The function is used to obtain all the opencl-enable devices
}
