% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{getJobStatus}
\alias{getJobStatus}
\title{Query the current job status in a device}
\usage{
getJobStatus(i)
}
\arguments{
\item{i}{integer The device index}
}
\description{
Query the current job status in a device
}
