% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgFunc.R
\name{gNumber}
\alias{gNumber}
\title{Create a scalar variable}
\usage{
gNumber(precision = GPUVar$default_float, constDef = FALSE)
}
\arguments{
\item{precision}{The variable type, please refer to \code{gpuMagic.getAvailableType()} to see the available data type.}

\item{constDef}{Specify if the variable can be redefined. The package will automatically update the variable definition when it is needed,
if you do not need this feature, you can manually turn the feature off.
It is useful in some special cases such as turning off the auto update to do the integer division
(By default, the package will convert the variable to the default float type before doing the division).}
}
\value{
a variable initialize with 0.
}
\description{
The function will create a scalar variable, it is only useful in the openCL functions.
It can also be called in R, but its argument will not take any effect.
}
\examples{
a=gNumber(precision='double',constDef=FALSE)

}
