% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCParserSupportFunc.R
\name{gpu_cast_int}
\alias{gpu_cast_int}
\title{Internal usage only, the package export this function only for the other package to access.}
\usage{
gpu_cast_int(x)
}
\arguments{
\item{x}{Internal usage only}
}
\description{
Internal usage only, the package export this function only for the other package to access.
}
\seealso{
Other internal cast functions: \code{\link{gpu_cast_double}},
  \code{\link{gpu_cast_float}},
  \code{\link{gpu_cast_long}}, \code{\link{gpu_cast_uint}},
  \code{\link{gpu_cast_ulong}}
}
\concept{internal cast functions}
