% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{gpuEmptMatrix}
\alias{gpuEmptMatrix}
\title{Create an empty matrix without initialization in an openCL device}
\usage{
gpuEmptMatrix(row = 1, col = 1, type = "auto", device = "auto")
}
\arguments{
\item{row, col}{the row and column number of the matrix}

\item{type}{The precision that is used to store the data.}

\item{device}{The device that the data is sent to.}
}
\description{
Create an empty matrix without initialization in an openCL device
}
