% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{gpuMagic.getAvailableType}
\alias{gpuMagic.getAvailableType}
\title{Get all the available openCL variable type}
\usage{
gpuMagic.getAvailableType()
}
\value{
A vector of all the available data type.
}
\description{
Get all the available openCL variable type
}
