% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{gpuMagic.getMemUsage}
\alias{gpuMagic.getMemUsage}
\title{Get the device memory usage}
\usage{
gpuMagic.getMemUsage()
}
\value{
No return value, the result will be printed in the console.
}
\description{
The function will print the memory usage on the console
}
\examples{
gpuMagic.getMemUsage()
}
