% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{gpuMagic.getOptions}
\alias{gpuMagic.getOptions}
\title{Get the openCL options}
\usage{
gpuMagic.getOptions(opt = "all")
}
\arguments{
\item{opt}{The options that the function will return. It can be either 'all' or a vector of the option names.}
}
\value{
A list of the options
}
\description{
The functions get the computing precision when compile the GPU code and the number of workers in a computing group.
}
\details{
The fields \code{default.float}, \code{default.int} and \code{default.index.type} are used to control the computing precision.
When transferring data from R to GPU, if the data in R has a numeric or double storage mode,
\code{default.float} will be used to convert data type.
Similarly, If the data has an Integer storage model. \code{default.int} will be used.

\code{default.index.type} controls the variable type for the for loop index, variable dimension etc.

\code{default.thread.num} is used to control the number of workers in a group in openCL. It is not expected to be changed unless you know what you are doing.
}
\examples{
#Get all the available options
opt=gpuMagic.getOptions()
opt
}
