% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{gpuMatrix}
\alias{gpuMatrix}
\title{Create a matrix in an openCL device}
\usage{
gpuMatrix(data, type = "auto", device = "auto")
}
\arguments{
\item{data}{It can be a matrix or an R object that can be converted into a matrix.}

\item{type}{The precision that is used to store the data.}

\item{device}{The device that the data is sent to.}
}
\description{
Create a matrix in an openCL device
}
