% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\docType{methods}
\name{length}
\alias{length}
\title{Length of an gpuMatrix object}
\usage{
\S4method{length}{gpuMatrix}(x)
}
\arguments{
\item{x}{an gpuMatrix object}
}
\description{
Get the length of an gpuMatrix object.
}
