% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgFunc.R
\name{Matrix}
\alias{Matrix}
\title{Create a matrix}
\usage{
Matrix(nrow = 1, ncol = 1, precision = GPUVar$default_float,
  constDef = FALSE, shared = FALSE, location = "global")
}
\arguments{
\item{nrow, ncol}{The matrix dimension.}

\item{precision}{The variable type, please refer to \code{gpuMagic.getAvailableType()} to see the available data type.}

\item{constDef}{Specify if the variable can be redefined. The package will automatically update the variable definition when it is needed,
if you do not need this feature, you can manually turn the feature off.
It is useful in some special cases such as turning off the auto update to do the integer division
(By default, the package will convert the variable to the default float type before doing the division).}

\item{shared}{If the matrix is shared by all the workers in a work group. Do not use it if you don't know its meaning.}

\item{location}{The physical memory location of the matrix, it can be either 'global' or 'local'. Do not use it if you don't know its meaning.}
}
\value{
a matrix initialize with 0.
}
\description{
The function create a matrix, it is only useful in the openCL functions.
it can also be called in R, but its argument may or may not take any effect.
}
\examples{
#Create a 10-by-10 matrix
A=Matrix(10,10)
}
