% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\docType{methods}
\name{ncol,gpuMatrix-method}
\alias{ncol,gpuMatrix-method}
\title{The Number of Rows/Columns of an gpuMatrix object}
\usage{
\S4method{ncol}{gpuMatrix}(x)
}
\arguments{
\item{x}{an gpuMatrix object}
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns present in \code{x}
}
