% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgFunc.R
\name{return_nocpy}
\alias{return_nocpy}
\alias{t_nocpy}
\title{No copy method}
\usage{
return_nocpy(x)

t_nocpy(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{return_nocpy}: No return value

\code{t_nocpy}: the transpose of \code{x}
}
\description{
Doing some opration without copying memory
}
\details{
\code{return_nocpy}: The usage of the \code{return_nocpy} is same as \code{return}. This feature is for openCL code only,
if it is called in R, the function \code{return()} will be called instead

\code{t_nocpy}: The function transposes \code{x} without allocating the memory. It only works for the openCL code,
if it is called in R, the function \code{t()} will be called instead
}
\examples{
x=matrix(0)
#return_nocpy(x)
#x=t_nocpy(x)
}
