% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{print.options}
\alias{print.options}
\title{Print the available options in a pretty format}
\usage{
\method{print}{options}(x, ...)
}
\arguments{
\item{x}{an options object.}

\item{...}{just for making the package checking happy.}
}
\value{
No return value, the result will be printed in the console
}
\description{
Print the available options in a pretty format
}
\examples{
opt=gpuMagic.getOptions()
#print.options(opt)
opt
}
