% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R, R/varInfoManager.R
\name{print.options}
\alias{print.options}
\alias{print.plainText}
\alias{print.varInfo}
\title{Print the available options in a pretty format}
\usage{
\method{print}{options}(x, ...)

\method{print}{plainText}(x, ...)

\method{print}{varInfo}(x, simplify = TRUE, printDef = FALSE, ...)
}
\arguments{
\item{x}{an options object.}

\item{...}{just for making the package checking happy.}

\item{simplify}{Specify whether only the important properties should be printed}

\item{printDef}{Whether the variable definition should be printed(version=0)}
}
\value{
No return value, the result will be printed in the console
}
\description{
Print the available options in a pretty format
}
\examples{
opt=gpuMagic.getOptions()
print(opt)
}
