% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgFunc.R
\name{return.nocpy}
\alias{return.nocpy}
\title{Return the result without memory copy}
\usage{
return.nocpy(x)
}
\arguments{
\item{x}{The return value}
}
\value{
No return value
}
\description{
The usage of the \code{return.nocpy} is same as \code{return}. This feature is for openCL code only, please not use it in R function.
}
