% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuFunctions.R
\name{setDevice}
\alias{setDevice}
\title{Set which device will be used in the opencl, call 'getDeviceList()' first to figure out the index before use this function}
\usage{
setDevice(i)
}
\arguments{
\item{i}{integer The device index}
}
\description{
Set which device will be used in the opencl, call 'getDeviceList()' first to figure out the index before use this function
}
