% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuMatix.R
\name{upload}
\alias{upload}
\title{Send the data to the device}
\usage{
upload(obj)
}
\arguments{
\item{obj}{an gpuMatrix object}
}
\description{
Send the data to the device
}
\details{
The function will automatically be called when an gpuMatrix object is created,
so the function is only needed when you want to update value of the matrix.
}
