% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelManager.R
\name{.kernel}
\alias{.kernel}
\title{Excute the openCL function}
\usage{
.kernel(src = "", kernel, parms, .device = "auto",
  .globalThreadNum = "length(FirstArg)", .options = kernel.getOption())
}
\value{
A vector or a matrix
}
\description{
Excute the openCL function
}
\examples{
#TODO
}
