% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuApply.R
\name{gpuSapply.getOption}
\alias{gpuSapply.getOption}
\title{Get the package compilation options}
\usage{
gpuSapply.getOption()
}
\description{
Get the package compilation options, the openCl compilation options(\code{kernel.getOption()}) are also included.
}
