% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelManager.R
\name{kernel.getOption}
\alias{kernel.getOption}
\title{Get the openCL compilation options}
\usage{
kernel.getOption()
}
\value{
A list of available options
}
\description{
Get the openCL compilation options
}
\examples{
opt=kernel.getOption()
opt
}
