% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgFunc.R
\name{return.nocpy}
\alias{return.nocpy}
\alias{t.nocpy}
\title{No copy method}
\usage{
return.nocpy(x)

t.nocpy(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{return.nocpy}: No return value

\code{t.nocpy}: the transpose of \code{x}
}
\description{
Doing some opration without memory copy
}
\details{
\code{return.nocpy}: The usage of the \code{return.nocpy} is same as \code{return}. This feature is for openCL code only,
if it is called in R, the function \code{return()} will be called instead

\code{t.nocpy}: The function transposes \code{x} without allocating the memory. It only works for the openCL code,
if it is called in R, the function \code{t()} will be called instead
}
\examples{
#return.nocpy(x)
#x=t.nocpy(x)
}
